# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Yet Another Node.js SugarApp Application API"
  },
  "host": "localhost:5000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/createuserprofile": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "This is User APi use to Create user profile",
        "description": "",
        "operationId": "updateUserProfile",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create user profile",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userfrofiles"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "definitions": {
...
     "userfrofiles": {
      "required": [
        "custID"
      ],
      "properties": {
        "custID": {
          "type": "string",
          "uniqueItems": true
        },
        "merchant": {
          "type":"string"
        },
        "food": {
          "type":"string"
        },
        "preferenceslocation": {
          "type":"string"
        },
        "phonenumber": {
          "type":"string"
        }
      }
    },
...
  }
}